+++ NEWS.md
@@ -3,3 +3,4 @@
  * Add targets functions.
  * Add functions to help with [CNAG](https://www.cnag.eu/) projects.
  * Add new function for transcriptomics.
+ * Add new function for file names.